created and moved helper functions to utils
