+++ recognition/ISIC_47044090/predict.py
@@ -4,7 +4,7 @@ import matplotlib.pyplot as plt
 import matplotlib.patches as patches
 import numpy as np
 from PIL import Image
-from dataset import scan_directory
+from utils import scan_directory
 
 modified_filepath = "./datasets/ISIC"
